Refactor RVIZ trajectory widget to allow it to be reused
